% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_sim.R
\name{dd_sim}
\alias{dd_sim}
\title{Function to simulate the diversity-dependent diversification process}
\usage{
dd_sim(pars, age, ddmodel = 1)
}
\description{
Function to simulate the diversity-dependent diversification process
}
\references{
\itemize{
\item Etienne, R.S. et al. 2012, Proc. Roy. Soc. B 279: 1300-1309,
doi: 10.1098/rspb.2011.1439 \cr - Etienne, R.S. & B. Haegeman 2012. Am. Nat.
180: E75-E89, doi: 10.1086/667574
}
}
\author{
Rampal S. Etienne
}
