% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{nn_bootstrap_uncertainty}
\alias{nn_bootstrap_uncertainty}
\title{Function to bootstrap the uncertainty of the EvoNN estimates}
\usage{
nn_bootstrap_uncertainty(estimate, scenario = "DDD", n = 100, timeout = 60)
}
\arguments{
\item{estimate}{The EvoNN estimate of the speciation rate, extinction rate, and carrying capacity (DDD)}

\item{scenario}{The diversification scenario: \cr \code{"BD"} : Birth-death
diversification \cr \code{"DDD"} : Diversity-dependent diversification}

\item{n}{The number of bootstrap iterations}

\item{timeout}{The maximum time in seconds for each iteration}
}
\value{
\item{ results_NN }{ A data frame with the following columns: \cr
\code{lambda} : Speciation rate \cr \code{mu} : Extinction rate \cr \code{cap} : Carrying capacity
(Only for DDD) \cr }
}
\description{
Generates posterior distributions of the speciation rate, extinction rate, and carrying capacity (DDD) of a
phylogenetic tree by bootstrapping the tree and estimating the parameters using EvoNN
}
\author{
Tianjian Qin
}
