% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{nn_estimate}
\alias{nn_estimate}
\title{Function to estimate phylogenetic parameters using EvoNN}
\usage{
nn_estimate(tree, scenario = "DDD")
}
\arguments{
\item{tree}{A phylogenetic tree in phylo format}

\item{scenario}{The diversification scenario: \cr \code{"BD"} : Birth-death
diversification \cr \code{"DDD"} : Diversity-dependent diversification}
}
\value{
\item{ out }{ A list with the following elements: \cr
\code{ pred_lambda } : Speciation rate \cr \code{ pred_mu } : Extinction rate \cr
\code{ pred_cap } : Carrying capacity (Only for DDD) \cr
\code{ scale } : A factor to scale the tree to the desired age (10) \cr
\code{ num_nodes } : Number of nodes in the tree \cr}
}
\description{
Estimates the speciation rate, extinction rate, and carrying capacity (DDD) of a
phylogenetic tree using the EvoNN pre-trained models
}
\author{
Tianjian Qin
}
