% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{nn_plot_bootstrap}
\alias{nn_plot_bootstrap}
\title{Function to plot the bootstrap results of the EvoNN estimates}
\usage{
nn_plot_bootstrap(bootstrap, result, scenario = "DDD")
}
\arguments{
\item{bootstrap}{The bootstrap results of the EvoNN estimates}

\item{result}{The EvoNN estimate of the speciation rate, extinction rate, and carrying capacity (DDD)}

\item{scenario}{The diversification scenario: \cr \code{"BD"} : Birth-death
diversification \cr \code{"DDD"} : Diversity-dependent diversification}
}
\value{
\item{ plots }{ A list of plots }
}
\description{
Plots the posterior distributions of the speciation rate, extinction rate, and carrying capacity (DDD) of a
phylogenetic tree estimated using EvoNN, with the original estimates
}
\author{
Tianjian Qin
}
